+++ docs/changelog.rst
@@ -21,6 +21,7 @@ Changed
 Fixed
 -----
 - ``ext_override`` option for Xarray Drivers is now properly used to filter uri's with a different extension.
+- ``expand_uri_paths`` function where the uri has no placeholders: ``{name}``, and only wildcards: ``*`` now correctly resolve.
 
 Deprecated
 ----------